Rename package etc. to pagerduty
