+++ .github/workflows/build.yml
@@ -34,6 +34,8 @@ jobs:
     if: github.ref == 'refs/heads/master'
     strategy:
       fail-fast: false
+      matrix:
+        node: [12]
 
     steps:
       - name: Setup Node
@@ -58,7 +60,6 @@ jobs:
           else
             echo "Version commit tag not found. Not publishing."
           fi
-
       - name: Publish
         if: env.publish == 'true'
         env:
